+++ src/main/java/pl/ug/virtualofficebackend/domain/decorationType/internal/DecorationTypeRepository.java
@@ -1,0 +1,9 @@
+package pl.ug.virtualofficebackend.domain.decorationType.internal;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+import pl.ug.virtualofficebackend.domain.decorationType.entity.DecorationType;
+
+@Repository
+public interface DecorationTypeRepository extends CrudRepository<DecorationType, Long> {
+}
