--- src/main/java/pl/ug/virtualofficebackend/domain/entity/Item.java
@@ -1,53 +1,0 @@
-package pl.ug.virtualofficebackend.domain.entity;
-
-import javax.persistence.*;
-
-@Entity(name = "item")
-public class Item {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    private String name;
-
-    @ManyToOne
-    private ItemType type;
-
-    @ManyToOne
-    private Workstation workstation;
-
-    //region GET SET
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long itemId) {
-        this.id = itemId;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public ItemType getType() {
-        return type;
-    }
-
-    public void setType(ItemType type) {
-        this.type = type;
-    }
-
-    public Workstation getWorkstation() {
-        return workstation;
-    }
-
-    public void setWorkstation(Workstation workstation) {
-        this.workstation = workstation;
-    }
-    //endregion
-
-}
