--- src/main/java/pl/ug/virtualofficebackend/domain/internal/RoomRepository.java
@@ -1,9 +1,0 @@
-package pl.ug.virtualofficebackend.domain.internal;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-import pl.ug.virtualofficebackend.domain.entity.Room;
-
-@Repository
-public interface RoomRepository extends CrudRepository<Room, Long> {
-}
