--- src/main/java/pl/ug/virtualofficebackend/domain/internal/RoomTypeRepository.java
@@ -1,9 +1,0 @@
-package pl.ug.virtualofficebackend.domain.internal;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-import pl.ug.virtualofficebackend.domain.entity.RoomType;
-
-@Repository
-public interface RoomTypeRepository extends CrudRepository<RoomType, Long> {
-}
