+++ src/main/java/pl/ug/virtualofficebackend/domain/itemType/internal/ItemTypeRepository.java
@@ -1,0 +1,9 @@
+package pl.ug.virtualofficebackend.domain.itemType.internal;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+import pl.ug.virtualofficebackend.domain.itemType.entity.ItemType;
+
+@Repository
+public interface ItemTypeRepository extends CrudRepository<ItemType, Long> {
+}
