+++ src/main/java/pl/ug/virtualofficebackend/domain/room/entity/Room.java
@@ -1,0 +1,110 @@
+package pl.ug.virtualofficebackend.domain.room.entity;
+
+import pl.ug.virtualofficebackend.domain.roomType.entity.RoomType;
+
+import javax.persistence.*;
+
+@Entity(name = "room")
+public class Room {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    private String name;
+    private int capacity;
+
+    /*
+        Na ten moment start i koniec przedmiotu
+        wszystko bedzie traktowane jako kwadrat
+
+        np. pokoj zaczyna sie w (0,0) i konczy w (1,1)
+
+        North/South/West/East
+     */
+
+    // (x1,y1) i (x2,y2)
+    private int x1Location;
+    private int y1Location;
+    private int x2Location;
+    private int y2Location;
+
+    private String rotation;
+
+    @ManyToOne
+    private RoomType type;
+
+    //region GET SET
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public int getCapacity() {
+        return capacity;
+    }
+
+    public void setCapacity(int capacity) {
+        this.capacity = capacity;
+    }
+
+    public int getX1Location() {
+        return x1Location;
+    }
+
+    public void setX1Location(int x1Location) {
+        this.x1Location = x1Location;
+    }
+
+    public int getY1Location() {
+        return y1Location;
+    }
+
+    public void setY1Location(int y1Location) {
+        this.y1Location = y1Location;
+    }
+
+    public int getX2Location() {
+        return x2Location;
+    }
+
+    public void setX2Location(int x2Location) {
+        this.x2Location = x2Location;
+    }
+
+    public int getY2Location() {
+        return y2Location;
+    }
+
+    public void setY2Location(int y2Location) {
+        this.y2Location = y2Location;
+    }
+
+    public String getRotation() {
+        return rotation;
+    }
+
+    public void setRotation(String rotation) {
+        this.rotation = rotation;
+    }
+
+    public RoomType getType() {
+        return type;
+    }
+
+    public void setType(RoomType type) {
+        this.type = type;
+    }
+    //endregion
+
+}
