+++ src/main/java/pl/ug/virtualofficebackend/domain/roomType/internal/RoomTypeRepository.java
@@ -1,0 +1,9 @@
+package pl.ug.virtualofficebackend.domain.roomType.internal;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+import pl.ug.virtualofficebackend.domain.roomType.entity.RoomType;
+
+@Repository
+public interface RoomTypeRepository extends CrudRepository<RoomType, Long> {
+}
